# tooling/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoinFlip PvP dApp API",
    "version": "1.0.0",
    "description": "API for the PvP CoinFlip decentralized application on Axiome Chain. Enables creating, accepting, revealing, and resolving Heads/Tails bets with LAUNCH tokens."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    },
    {
      "url": "https://api.coinflip.example.com",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/v1/bets": {
      "get": {
        "tags": [
          "Bets"
        ],
        "summary": "List open bets",
        "operationId": "getBets",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "accepting",
                "accepted",
                "revealed",
                "canceled",
                "timeout_claimed"
              ]
            }
          },
          {
            "name": "min_amount",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "max_amount",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of bets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Bet"
                      }
                    },
                    "cursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "has_more": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "cursor",
                    "has_more"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Bets"
        ],
        "summary": "Create a new bet",
        "operationId": "createBet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bet created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Bet"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vault/deposit": {
      "post": {
        "tags": [
          "Vault"
        ],
        "summary": "Initiate deposit (returns Axiome Connect payload)",
        "operationId": "depositToVault",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deposit payload for Axiome Connect"
          }
        }
      }
    },
    "/api/v1/vault/withdraw": {
      "post": {
        "tags": [
          "Vault"
        ],
        "summary": "Withdraw LAUNCH from vault",
        "operationId": "withdrawFromVault",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Withdraw initiated"
          }
        }
      }
    },
...
    "/api/v1/users/{address}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get public user profile",
        "operationId": "getUserByAddress",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserProfile"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leaderboard": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get leaderboard",
        "operationId": "getLeaderboard",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LeaderboardEntry"
                      }
                    },
                    "cursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "has_more": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "cursor",
                    "has_more"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Bet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "description": "Unique bet identifier",
            "example": 42
          },
          "maker": {
            "type": "string",
            "minLength": 1,
            "description": "Axiome wallet address",
            "example": "axm1abc123..."
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d+$",
            "description": "Token amount as string",
            "example": "100"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "accepting",
              "accepted",
              "revealed",
              "canceled",
              "timeout_claimed"
            ],
            "description": "Current bet status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Bet creation time (ISO 8601)"
          },
          "txhash_create": {
            "type": "string",
            "description": "Transaction hash",
            "example": "A1B2C3D4E5F6..."
          },
          "acceptor": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "description": "Acceptor address (null if OPEN)",
            "example": "axm1abc123..."
          },
          "acceptor_guess": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "heads",
              "tails",
              null
            ],
            "description": "Acceptor guess (null if OPEN)"
          },
          "accepted_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "txhash_accept": {
            "type": [
              "string",
              "null"
            ],
            "description": "Transaction hash",
            "example": "A1B2C3D4E5F6..."
          },
          "reveal_side": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "heads",
              "tails",
              null
            ],
            "description": "Maker revealed side"
          },
          "winner": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "description": "Winner address",
            "example": "axm1abc123..."
          },
          "payout_amount": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^\\d+$",
            "description": "Winner payout (after commission)",
            "example": "100"
          },
          "commission_amount": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^\\d+$",
            "description": "Token amount as string",
            "example": "100"
          },
          "resolved_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "txhash_resolve": {
            "type": [
              "string",
              "null"
            ],
            "description": "Transaction hash",
            "example": "A1B2C3D4E5F6..."
          },
          "reveal_deadline": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Deadline for maker to reveal (accepted_at + 5min)"
          }
        },
        "required": [
          "id",
          "maker",
          "amount",
          "status",
          "created_at",
          "txhash_create",
          "acceptor",
          "acceptor_guess",
          "accepted_at",
          "txhash_accept",
          "reveal_side",
          "winner",
          "payout_amount",
          "commission_amount",
          "resolved_at",
          "txhash_resolve",
          "reveal_deadline"
        ]
      },
      "CreateBetRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string",
            "pattern": "^\\d+$",
            "description": "Bet amount in LAUNCH tokens",
            "example": "100"
          }
        },
        "required": [
          "amount"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code",
                "example": "INSUFFICIENT_BALANCE"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "additionalProperties": {},
                "description": "Additional error details"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "DepositRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string",
            "pattern": "^\\d+$",
            "description": "Amount of LAUNCH to deposit",
            "example": "500"
          }
        },
        "required": [
          "amount"
        ]
      },
      "WithdrawRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string",
            "pattern": "^\\d+$",
            "description": "Amount of LAUNCH to withdraw",
            "example": "200"
          }
        },
        "required": [
          "amount"
        ]
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "minLength": 1,
            "description": "Axiome wallet address",
            "example": "axm1abc123..."
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name"
          },
          "avatar_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri",
            "description": "Avatar image URL"
          },
          "stats": {
            "type": "object",
            "properties": {
              "total_bets": {
                "type": "integer",
                "description": "Total bets placed",
                "example": 42
              },
              "wins": {
                "type": "integer",
                "description": "Total wins",
                "example": 25
              },
              "losses": {
                "type": "integer",
                "description": "Total losses",
                "example": 17
              },
              "total_wagered": {
                "type": "string",
                "pattern": "^\\d+$",
                "description": "Total LAUNCH wagered",
                "example": "5000"
              },
              "total_won": {
                "type": "string",
                "pattern": "^\\d+$",
                "description": "Total LAUNCH won",
                "example": "3200"
              }
            },
            "required": [
              "total_bets",
              "wins",
              "losses",
              "total_wagered",
              "total_won"
            ]
          },
          "authz_enabled": {
            "type": "boolean",
            "description": "1-click mode active"
          },
          "authz_expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Authz grant expiration"
          },
          "fee_sponsored": {
            "type": "boolean",
            "description": "Gas sponsorship active"
          }
        },
        "required": [
          "address",
          "nickname",
          "avatar_url",
          "stats",
          "authz_enabled",
          "authz_expires_at",
          "fee_sponsored"
        ]
      },
      "LeaderboardEntry": {
        "type": "object",
        "properties": {
          "rank": {
            "type": "integer",
            "example": 1
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "description": "Axiome wallet address",
            "example": "axm1abc123..."
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ]
          },
          "wins": {
            "type": "integer"
          },
          "total_wagered": {
            "type": "string",
            "pattern": "^\\d+$",
            "description": "Token amount as string",
            "example": "100"
          },
          "win_rate": {
            "type": "number",
            "description": "Win rate 0-1",
            "example": 0.65
          }
        },
        "required": [
          "rank",
          "address",
          "nickname",
          "wins",
          "total_wagered",
          "win_rate"
        ]
      },
...
    }
  }
}